Dedup line models and mailer locale switching
Extract LineItem concern with TYPE_OPTIONS, inheritance_column, is_item?,
and title/type validations. InvoiceLine and DeliveryNoteLine include it.

Lift I18n.with_locale(customer.language.iso_code) into
ApplicationMailer#with_customer_locale; used by InvoiceMailer and
DeliveryNoteMailer.
